requirements.txt
google-generativeai
streamlit
biopython